+++ src/blueprints.py
@@ -1,4 +1,7 @@
+from flask import Blueprint as FlBlueprint
 from flask_smorest import Blueprint
+
+__all__ = ["all_blueprints", "route_embed"]
 
 
 def _factory(partial_module_string: str, url_prefix: str) -> Blueprint:
@@ -11,3 +14,4 @@ root = _factory("root", "/")
 linkrot = _factory("linkrot", "/linkrot")
 
 all_blueprints = (root, linkrot)
+route_embed = FlBlueprint("embed", "src.views.embed", url_prefix="/webring-embed.js")